requirements.txt
fastapi
uvicorn
python-dotenv
opencv-python
numpy
pillow
httpx